// reit-portal/src/app/models/listing-detail.ts
export interface IListingDetail {
  imageURL?: string;
  listingPrice?: number;
  numOfBedrooms?: number;
  numOfBathrooms?: number;
  numOfStories?: number;
  numOfLevels?: number;
  numOfParkingSpaces?: number;
  lotSize?: string;
  numOfGarageSpaces?: number;
  homeType?: string;
  propertyCondition?: string;
  yearBuilt?: number;
  hasHOA?: boolean;
  hoaFee?: number;
  propertyAge?: number;
}